feat: Extract createElementFromHtml function in separate module
