+++ js/createElementFromHtml.js
@@ -1,0 +1,8 @@
+const createElementFromHtml = (html) => {
+  const template = document.createElement(`template`);
+  html = html.trim();
+  template.innerHTML = html;
+  return template.content.firstChild;
+};
+
+export default createElementFromHtml;
